specified Errors/http codes

+++ nestjs-api-webapp/src/hardware/hardware.service.ts
@@ -1,4 +1,4 @@
-import {Injectable} from '@nestjs/common';
+import {Injectable, NotFoundException} from '@nestjs/common';
 import {PrismaService} from '../prisma/prisma.service';
 import {HardwareDto} from './dto';
 import {Hardware} from '@prisma/client';
@@ -24,37 +24,62 @@ export class HardwareService {
 	}
 
 	async getOne(id: number): Promise<Hardware> {
-		return this.prisma.hardware.findUnique({
-			where: {
-				id: id,
-			},
-		});
-	}
-
-	async update(id: number, dto: HardwareDto): Promise<Hardware> {
 		try {
-			const hardware = await this.prisma.hardware.update({
+			const hardware = await this.prisma.hardware.findUnique({
 				where: {
 					id: id,
 				},
-				data: dto,
 			});
-
+			if (!hardware) {
+				// Return appropriate HTTP code (e.g., 404 Not Found) if hardware is not found
+				throw new NotFoundException('Hardware not found');
+			}
 			return hardware;
 		} catch (error) {
 			throw error;
 		}
 	}
 
-	async delete(id: number): Promise<Hardware> {
+	async update(id: number, dto: HardwareDto): Promise<Hardware> {
 		try {
-			const hardware = await this.prisma.hardware.delete({
+			const hardware = await this.prisma.hardware.findUnique({
+				where: {
+					id: id,
+				},
+			});
+			if (!hardware) {
+				throw new NotFoundException('Hardware not found');
+			}
+			const updatedHardware = await this.prisma.hardware.update({
+				where: {
+					id: id,
+				},
+				data: dto,
+			});
+			return updatedHardware;
+		} catch (error) {
+			throw error;
+		}
+	}
+
+	async delete(id: number): Promise<object> {
+		try {
+			const hardware = await this.prisma.hardware.findUnique({
+				where: {
+					id: id,
+				},
+			});
+			if (!hardware) {
+				throw new NotFoundException('Hardware not found');
+			}
+
+			await this.prisma.hardware.delete({
 				where: {
 					id: id,
 				},
 			});
 
-			return hardware;
+			return {message: 'Hardware deleted'};
 		} catch (error) {
 			throw error;
 		}
